chapter2-02/pyproject.toml
[project]
name = "chapter2-02"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.99.9",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
